refactor(api): move SiteExplorerConfig into site_explorer module
Extract SiteExplorerConfig (and SiteExplorerExploreMode) from the
3000-line cfg/file.rs god-file into a new sibling module
site_explorer/config.rs. The config now lives next to the code it
configures — a prerequisite for eventually extracting site_explorer
into its own crate.

The bmc_proxy-specific serde helpers stay with SiteExplorerConfig since
they're specific to that field. Generic serde helpers used across many
config structs (deserialize_arc_atomic_bool, serialize_arc_atomic_bool,
as_duration, as_std_duration) are moved to a new carbide-utils::config
module so other config types can share them.

`pub use site_explorer::config::SiteExplorerExploreMode` in lib.rs
preserves the existing re-export. No behavioral changes.

Part of the effort to break up the carbide-api mega-crate.
