+++ crates/api/src/site_explorer/config.rs
@@ -1,0 +1,340 @@
+/*
+ * SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+use std::ops::Deref;
+use std::sync::Arc;
+use std::sync::atomic::{AtomicBool, Ordering as AtomicOrdering};
+
+use arc_swap::ArcSwap;
+use chrono::Duration;
+use duration_str::{deserialize_duration, deserialize_duration_chrono};
+use serde::{Deserialize, Deserializer, Serialize, Serializer};
+use utils::HostPortPair;
+use utils::config::{
+    as_duration, as_std_duration, deserialize_arc_atomic_bool, serialize_arc_atomic_bool,
+};
+
+/// SiteExplorer related configuration for hardware discovery and ingestion.
+#[derive(Clone, Debug, Deserialize, Serialize)]
+pub struct SiteExplorerConfig {
+    /// Whether SiteExplorer is enabled. Dynamically toggleable at runtime via SetDynamicConfig.
+    #[serde(
+        default = "SiteExplorerConfig::default_enabled",
+        deserialize_with = "deserialize_arc_atomic_bool",
+        serialize_with = "serialize_arc_atomic_bool"
+    )]
+    pub enabled: Arc<AtomicBool>,
+    /// The interval at which site explorer runs.
+    /// Defaults to 5 Minutes if not specified.
+    #[serde(
+        default = "SiteExplorerConfig::default_run_interval",
+        deserialize_with = "deserialize_duration",
+        serialize_with = "as_std_duration"
+    )]
+    pub run_interval: std::time::Duration,
+    /// The maximum amount of nodes that are explored concurrently.
+    /// Default is 5.
+    #[serde(default = "SiteExplorerConfig::default_concurrent_explorations")]
+    pub concurrent_explorations: u64,
+    /// How many nodes should be explored in a single run.
+    /// Default is 10.
+    /// This number divided by `concurrent_explorations` will determine how many
+    /// exploration batches are needed inside a run.
+    /// If the value is set too high the site exploration will take a lot of time
+    /// and the exploration report will be updated less frequent. Therefore it
+    /// is recommended to reduce `run_interval` instead of increasing
+    /// `explorations_per_run`.
+    #[serde(default = "SiteExplorerConfig::default_explorations_per_run")]
+    pub explorations_per_run: u64,
+
+    /// Whether SiteExplorer should create Managed Host state machine
+    #[serde(
+        default = "SiteExplorerConfig::default_create_machines",
+        deserialize_with = "deserialize_arc_atomic_bool",
+        serialize_with = "serialize_arc_atomic_bool"
+    )]
+    pub create_machines: Arc<AtomicBool>,
+
+    /// How many ManagedHosts should be created in a single run. Default is 4.
+    #[serde(default = "SiteExplorerConfig::default_machines_created_per_run")]
+    pub machines_created_per_run: u64,
+
+    /// Whether SiteExplorer should rotate/update Switch NVOS admin credentials
+    #[serde(
+        default = "SiteExplorerConfig::default_rotate_switch_nvos_credentials",
+        deserialize_with = "deserialize_arc_atomic_bool",
+        serialize_with = "serialize_arc_atomic_bool"
+    )]
+    pub rotate_switch_nvos_credentials: Arc<AtomicBool>,
+
+    /// DEPRECATED: Use `bmc_proxy` instead.
+    /// The IP address to connect to instead of the BMC that made the dhcp request.
+    /// This is a debug override and should not be used in production.
+    pub override_target_ip: Option<String>,
+
+    /// DEPRECATED: Use `bmc_proxy` instead.
+    /// The port to connect to for redfish requests.
+    /// This is a debug override and should not be used in production.
+    pub override_target_port: Option<u16>,
+
+    /// Whether to allow hosts with zero DPUs in site-explorer. This should typically be set to
+    /// false in production environments where we expect all hosts to have DPUs. When false, if we
+    /// encounter a host with no DPUs, site-explorer will throw an error for that host (because it
+    /// should be assumed that there's a bug in detecting the DPUs).
+    #[serde(default)]
+    pub allow_zero_dpu_hosts: bool,
+
+    /// The host:port to use as a proxy when making BMC calls to all hosts in NICo. This is used
+    /// for integration testing, and for local development with machine-a-tron/bmc-mock. Should not
+    /// be used in production.
+    #[serde(
+        default,
+        deserialize_with = "deserialize_bmc_proxy",
+        serialize_with = "serialize_bmc_proxy"
+    )]
+    pub bmc_proxy: Arc<ArcSwap<Option<HostPortPair>>>,
+
+    /// If set to `true`, the server will allow changes to the `bmc_proxy` setting at runtime.
+    /// Defaults to true if the server is launched with `bmc_proxy` set, false otherwise.
+    /// If explicitly set to true or false, that value is respected for the lifetime of the process.
+    #[serde(default)]
+    pub allow_changing_bmc_proxy: Option<bool>,
+
+    /// Minimum time between consecutive force-restarts or BMC resets initiated by SiteExplorer.
+    /// Default is 1 hour.
+    #[serde(
+        default = "SiteExplorerConfig::default_reset_rate_limit",
+        deserialize_with = "deserialize_duration_chrono",
+        serialize_with = "as_duration"
+    )]
+    pub reset_rate_limit: Duration,
+
+    /// When true, non-DPU hosts use the `HostInband` admin network segment type instead of `Admin`.
+    #[serde(
+        default = "SiteExplorerConfig::default_admin_segment_type_non_dpu",
+        deserialize_with = "deserialize_arc_atomic_bool",
+        serialize_with = "serialize_arc_atomic_bool"
+    )]
+    pub admin_segment_type_non_dpu: Arc<AtomicBool>,
+
+    /// Whether site-controller should allocate a secondary
+    /// VTEP IP or leave that to discovery.
+    /// Current secondary VTEP use-case is additional
+    /// VTEP IPs for GENEVE VTEPS (GTEPS) used by traffic-intercept users.
+    ///  Only sites expected to support
+    /// additional VTEPS would turn this on.
+    #[serde(default)]
+    pub allocate_secondary_vtep_ip: bool,
+
+    /// Whether SiteExplorer should create Power Shelf state machine
+    #[serde(
+        default = "SiteExplorerConfig::default_create_power_shelves",
+        deserialize_with = "deserialize_arc_atomic_bool",
+        serialize_with = "serialize_arc_atomic_bool"
+    )]
+    pub create_power_shelves: Arc<AtomicBool>,
+
+    /// Whether SiteExplorer should create Power Shelf state machine from static IP
+    #[serde(
+        default = "SiteExplorerConfig::default_explore_power_shelves_from_static_ip",
+        deserialize_with = "deserialize_arc_atomic_bool",
+        serialize_with = "serialize_arc_atomic_bool"
+    )]
+    pub explore_power_shelves_from_static_ip: Arc<AtomicBool>,
+
+    /// How many Power Shelves should be created in a single run.
+    /// Default is 1.
+    #[serde(default = "SiteExplorerConfig::default_power_shelves_created_per_run")]
+    pub power_shelves_created_per_run: u64,
+
+    /// Whether SiteExplorer should create Switch state machine
+    #[serde(
+        default = "SiteExplorerConfig::default_create_switches",
+        deserialize_with = "deserialize_arc_atomic_bool",
+        serialize_with = "serialize_arc_atomic_bool"
+    )]
+    pub create_switches: Arc<AtomicBool>,
+
+    /// How many Switches should be created in a single run.
+    /// Default is 9.
+    #[serde(default = "SiteExplorerConfig::default_switches_created_per_run")]
+    pub switches_created_per_run: u64,
+
+    /// Use onboard NIC for host networking instead of DPU NICs.
+    #[serde(
+        default = "SiteExplorerConfig::default_force_dpu_nic_mode",
+        deserialize_with = "deserialize_arc_atomic_bool",
+        serialize_with = "serialize_arc_atomic_bool"
+    )]
+    pub force_dpu_nic_mode: Arc<AtomicBool>,
+    /// Controls which Redfish client implementation is used
+    /// for hardware discovery (LibRedfish, NvRedfish, or
+    /// CompareResult for side-by-side validation).
+    #[serde(default = "SiteExplorerConfig::default_explore_mode")]
+    pub explore_mode: SiteExplorerExploreMode,
+}
+
+impl Default for SiteExplorerConfig {
+    fn default() -> Self {
+        SiteExplorerConfig {
+            enabled: Arc::new(true.into()),
+            run_interval: Self::default_run_interval(),
+            concurrent_explorations: Self::default_concurrent_explorations(),
+            explorations_per_run: Self::default_explorations_per_run(),
+            create_machines: Arc::new(true.into()),
+            machines_created_per_run: Self::default_machines_created_per_run(),
+            override_target_ip: None,
+            override_target_port: None,
+            allow_zero_dpu_hosts: false,
+            bmc_proxy: bmc_proxy(None),
+            allow_changing_bmc_proxy: None,
+            reset_rate_limit: Self::default_reset_rate_limit(),
+            admin_segment_type_non_dpu: Self::default_admin_segment_type_non_dpu(),
+            allocate_secondary_vtep_ip: false,
+            create_power_shelves: Arc::new(true.into()),
+            explore_power_shelves_from_static_ip: Arc::new(true.into()),
+            power_shelves_created_per_run: Self::default_power_shelves_created_per_run(),
+            create_switches: Arc::new(true.into()),
+            switches_created_per_run: Self::default_switches_created_per_run(),
+            rotate_switch_nvos_credentials: Self::default_rotate_switch_nvos_credentials(),
+            force_dpu_nic_mode: Arc::new(false.into()),
+            explore_mode: Self::default_explore_mode(),
+        }
+    }
+}
+
+impl PartialEq for SiteExplorerConfig {
+    fn eq(&self, other: &SiteExplorerConfig) -> bool {
+        self.enabled.load(AtomicOrdering::Relaxed) == other.enabled.load(AtomicOrdering::Relaxed)
+            && self.run_interval == other.run_interval
+            && self.concurrent_explorations == other.concurrent_explorations
+            && self.explorations_per_run == other.explorations_per_run
+            && self.create_machines.load(AtomicOrdering::Relaxed)
+                == other.create_machines.load(AtomicOrdering::Relaxed)
+            && self.override_target_ip == other.override_target_ip
+            && self.override_target_port == other.override_target_port
+    }
+}
+
+impl SiteExplorerConfig {
+    pub const fn default_run_interval() -> std::time::Duration {
+        std::time::Duration::from_secs(120)
+    }
+
+    pub fn default_enabled() -> Arc<AtomicBool> {
+        Arc::new(true.into())
+    }
+
+    pub fn default_create_machines() -> Arc<AtomicBool> {
+        Arc::new(true.into())
+    }
+
+    pub const fn default_concurrent_explorations() -> u64 {
+        30
+    }
+
+    pub const fn default_explorations_per_run() -> u64 {
+        90
+    }
+
+    pub const fn default_machines_created_per_run() -> u64 {
+        4
+    }
+
+    pub fn default_rotate_switch_nvos_credentials() -> Arc<AtomicBool> {
+        Arc::new(false.into())
+    }
+
+    pub const fn default_reset_rate_limit() -> Duration {
+        Duration::hours(1)
+    }
+
+    pub fn default_admin_segment_type_non_dpu() -> Arc<AtomicBool> {
+        Arc::new(false.into())
+    }
+
+    pub fn default_create_power_shelves() -> Arc<AtomicBool> {
+        Arc::new(false.into())
+    }
+
+    pub fn default_explore_power_shelves_from_static_ip() -> Arc<AtomicBool> {
+        Arc::new(false.into())
+    }
+
+    pub const fn default_power_shelves_created_per_run() -> u64 {
+        1
+    }
+
+    pub fn default_create_switches() -> Arc<AtomicBool> {
+        Arc::new(false.into())
+    }
+
+    pub const fn default_switches_created_per_run() -> u64 {
+        9
+    }
+
+    pub fn default_force_dpu_nic_mode() -> Arc<AtomicBool> {
+        Arc::new(false.into())
+    }
+
+    pub const fn default_explore_mode() -> SiteExplorerExploreMode {
+        SiteExplorerExploreMode::LibRedfish
+    }
+}
+
+pub fn bmc_proxy(s: Option<HostPortPair>) -> Arc<ArcSwap<Option<HostPortPair>>> {
+    Arc::new(ArcSwap::new(Arc::new(s)))
+}
+
+/// Selects the Redfish client backend used by SiteExplorer
+/// for BMC discovery.
+#[derive(Clone, Copy, Debug, Deserialize, Serialize)]
+pub enum SiteExplorerExploreMode {
+    /// Use the libredfish Rust client.
+    #[serde(rename = "libredfish")]
+    LibRedfish,
+    /// Use the NVIDIA-specific Redfish client.
+    #[serde(rename = "nv-redfish")]
+    NvRedfish,
+    /// Run both clients and compare results for validation.
+    #[serde(rename = "compare-result")]
+    CompareResult,
+}
+
+pub fn deserialize_bmc_proxy<'de, D>(
+    deserializer: D,
+) -> Result<Arc<ArcSwap<Option<HostPortPair>>>, D::Error>
+where
+    D: Deserializer<'de>,
+{
+    let p = Option::deserialize(deserializer)?;
+    Ok(Arc::new(ArcSwap::new(Arc::new(p))))
+}
+
+pub fn serialize_bmc_proxy<S>(
+    val: &Arc<ArcSwap<Option<HostPortPair>>>,
+    s: S,
+) -> Result<S::Ok, S::Error>
+where
+    S: Serializer,
+{
+    if let Some(val) = val.load().deref().deref() {
+        s.serialize_str(val.to_string().as_str())
+    } else {
+        s.serialize_none()
+    }
+}
